backend/requirements.txt
# Math Tutor Backend - Python 3.9 Compatible

# Core FastAPI dependencies
fastapi==0.116.0
uvicorn==0.35.0
python-multipart==0.0.20

# Environment and configuration
python-dotenv==1.1.1

# Google AI/Gemini
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0

# LangChain
langchain==0.3.26
langchain-core==0.3.68
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4

# Math and visualization
matplotlib==3.9.4
numpy==2.0.2
pillow==11.3.0

# HTTP and utilities
requests==2.32.4
requests-toolbelt==1.0.0
httpx==0.28.1
httpcore==1.0.9

# Data handling
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
SQLAlchemy==2.0.41
orjson==3.10.18
ormsgpack==1.10.0

# Utilities
tenacity==9.1.2
tqdm==4.67.1
PyYAML==6.0.2
click==8.1.8
jsonpatch==1.33
jsonpointer==3.0.0
packaging==24.2

# System dependencies
certifi>=2025.6.15
charset-normalizer==3.4.2
urllib3==2.5.0
idna==3.10
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
starlette==0.46.2
typing-extensions==4.14.1
annotated-types==0.7.0
exceptiongroup==1.3.0

# Additional dependencies
async-timeout==4.0.3
cachetools==5.5.2
filetype==1.2.0
fonttools==4.58.5
contourpy==1.3.0
cycler==0.12.1
kiwisolver==1.4.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
importlib-resources==6.5.2
zipp==3.23.0
zstandard==0.23.0
xxhash==3.5.0
six==1.17.0

# Protocol buffers and gRPC
protobuf==5.29.5
grpcio==1.73.1
grpcio-status==1.71.2
googleapis-common-protos==1.70.0
proto-plus==1.26.1

# Authentication
pyasn1==0.6.1
pyasn1-modules==0.4.2
rsa==4.9.1

# HTTP libraries
httplib2==0.22.0
uritemplate==4.2.0

# Type checking
typing-inspection==0.4.1
